Move single-nft example to an integration test
