+++ x/programs/rust/examples/single_nft/tests/example.rs
@@ -1,7 +1,7 @@
+use serial_test::serial;
 use std::env;
-
 use wasmlanche_sdk::simulator::{
-    id_from_step, Client, Endpoint, Key, Param, ParamType, Plan, PlanResponse, Step,
+    self, id_from_step, Endpoint, Key, Param, ParamType, Plan, PlanResponse, Step,
 };
 
 pub fn initialize_plan(
@@ -110,8 +110,24 @@ pub fn initialize_plan(
     }
 }
 
-#[allow(dead_code)]
-fn run() -> Result<Vec<PlanResponse>, Box<dyn std::error::Error>> {
+// export SIMULATOR_PATH=/path/to/simulator
+// export PROGRAM_PATH=/path/to/program.wasm
+#[test]
+#[serial]
+#[ignore = "requires SIMULATOR_PATH and PROGRAM_PATH to be set"]
+fn test_single_nft_plan() {
+    use wasmlanche_sdk::simulator::{self, Key};
+    let s_path = env::var(simulator::PATH_KEY).expect("SIMULATOR_PATH not set");
+    let simulator = simulator::Client::new(s_path);
+
+    let alice_key = "alice_key";
+    // create owner key in single step
+    let resp = simulator
+        .key_create::<PlanResponse>(alice_key, Key::Ed25519)
+        .unwrap();
+    assert_eq!(resp.error, None);
+
+    // create multiple step test plan
     let nft_name = "MyNFT".to_string();
     let binding = nft_name.len().to_string();
     let nft_name_length = binding.to_string();
@@ -132,8 +148,39 @@ fn run() -> Result<Vec<PlanResponse>, Box<dyn std::error::Error>> {
         nft_uri,
         nft_uri_length,
     );
-    let client = Client::new("path to simulator".to_owned());
 
-    let tx = client.run::<PlanResponse>(&plan)?;
-    Ok(tx)
+    // run plan
+    let plan_responses = simulator.run::<PlanResponse>(&plan).unwrap();
+
+    assert!(
+        plan_responses.iter().all(|resp| resp.error.is_none()),
+        "error: {:?}",
+        plan_responses
+            .iter()
+            .filter_map(|resp| resp.error.as_ref())
+            .next()
+    );
+}
+
+#[test]
+#[serial]
+#[ignore = "requires SIMULATOR_PATH and PROGRAM_PATH to be set"]
+fn test_create_nft_program() {
+    let s_path = env::var(simulator::PATH_KEY).expect("SIMULATOR_PATH not set");
+    let simulator = simulator::Client::new(s_path);
+
+    let alice_key = "alice_key";
+    // create alice key in single step
+    let resp = simulator
+        .key_create::<PlanResponse>(alice_key, Key::Ed25519)
+        .unwrap();
+    assert_eq!(resp.error, None);
+
+    let p_path = env::var("PROGRAM_PATH").expect("PROGRAM_PATH not set");
+    // create a new program on chain.
+    let resp = simulator
+        .program_create::<PlanResponse>("owner", p_path.as_ref())
+        .unwrap();
+    assert_eq!(resp.error, None);
+    assert!(resp.result.id.is_some());
 }
